+++ R/helpers_gghistostats_graphics.R
@@ -30,32 +30,28 @@
 # function body
 histo_labeller <- function(plot,
                            x,
-                           type = "parametric",
-                           tr = 0.2,
-                           k = 2L,
-                           centrality.line.args = list(color = "blue", size = 1),
+                           centrality.line.args = list(
+                             color = "blue",
+                             size = 1,
+                             linetype = "dashed"
+                           ),
                            ...) {
-
-  # create a temporary dataframe
-  dat_temp <- data.frame(.temp = ".temp", "var" = x)
-
-  # compute centrality measure
-  centrality_df <- centrality_data(dat_temp, .temp, var, type = type, tr = tr, k = k)
+  # compute centrality measure (with a temporary dataframe)
+  df_central <- centrality_data(data.frame(.x = ".x", "var" = x), .x, var, ...)
 
   # adding a vertical line corresponding to centrality parameter
   plot +
     rlang::exec(
-      .f = ggplot2::geom_vline,
-      xintercept = centrality_df$var[[1]],
-      !!!centrality.line.args,
-      linetype = "dashed"
+      ggplot2::geom_vline,
+      xintercept = df_central$var[[1]],
+      !!!centrality.line.args
     ) +
     ggplot2::scale_x_continuous(
       sec.axis = ggplot2::sec_axis(
         trans = ~.,
         name = NULL,
-        labels = parse(text = centrality_df$label[[1]]),
-        breaks = centrality_df$var[[1]]
+        labels = parse(text = df_central$label[[1]]),
+        breaks = df_central$var[[1]]
       )
     )
 }